+++ addons/im_livechat/models/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*
+
 from . import res_users
 from . import res_partner
 from . import im_livechat_channel
 from . import mail_channel
+from . import mail_channel_partner
 from . import rating
 from . import digest
