+++ addons/im_livechat/models/mail_channel.py
@@ -3,24 +3,6 @@
 
 from odoo import api, fields, models, _
 
-class ChannelPartner(models.Model):
-    _inherit = 'mail.channel.partner'
-
-    @api.autovacuum
-    def _gc_unpin_livechat_sessions(self):
-        """ Unpin livechat sessions with no activity for at least one day to
-            clean the operator's interface """
-        self.env.cr.execute("""
-            UPDATE mail_channel_partner
-            SET is_pinned = false
-            WHERE id in (
-                SELECT cp.id FROM mail_channel_partner cp
-                INNER JOIN mail_channel c on c.id = cp.channel_id
-                WHERE c.channel_type = 'livechat' AND cp.is_pinned is true AND
-                    cp.write_date < current_timestamp - interval '1 day'
-            )
-        """)
-
 
 class MailChannel(models.Model):
     """ Chat Session
