+++ addons/mail/models/mail_channel.py
@@ -49,9 +49,15 @@ class Channel(models.Model):
     email_send = fields.Boolean('Send messages by email', default=False)
     image_128 = fields.Image("Image", max_width=128, max_height=128, default=_get_default_image)
     # members (depends=['...'] is for `test_mail/tests/common.py`, class Moderation, `setUpClass`)
-    channel_last_seen_partner_ids = fields.One2many('mail.channel.partner', 'channel_id', string='Last Seen', depends=['channel_partner_ids'])
-    channel_partner_ids = fields.Many2many('res.partner', 'mail_channel_partner', 'channel_id', 'partner_id', string='Listeners', depends=['channel_last_seen_partner_ids'])
-    channel_message_ids = fields.Many2many('mail.message', 'mail_message_mail_channel_rel')
+    channel_partner_ids = fields.Many2many(
+        'res.partner', 'mail_channel_partner', 'channel_id', 'partner_id',
+        string='Listeners', depends=['channel_last_seen_partner_ids'])
+    channel_last_seen_partner_ids = fields.One2many(
+        'mail.channel.partner', 'channel_id', string='Last Seen',
+        depends=['channel_partner_ids'])
+    channel_message_ids = fields.Many2many(
+        'mail.message', 'mail_message_mail_channel_rel',
+        string='Channel Messages')
     is_member = fields.Boolean('Is a member', compute='_compute_is_member')
     is_subscribed = fields.Boolean(
         'Is Subscribed', compute='_compute_is_subscribed')
